crowdfunding/requirements.txt
beautifulsoup4==4.9.0
Django==1.11.29
django-crispy-forms==1.6.0
django-reversion==1.7.1
django-paypal==0.3.6
django-pagseguro==1.4.2
django-pure-pagination==0.3.0
django-dropbox==0.0.2
django-feedly==0.6.8
django-ship==0.2.7
django-socialize==0.5.4
django-distill==1.7
future==0.18.2
pypugjs==5.9.4
gdata-python3==3.0.1
gunicorn==20.0.4
httplib2==0.8
oauth==1.0.1
oauth2==1.9.0.post1
psycopg2==2.8.5
requests==2.23.0
# wsgiref==0.1.2
